requirements.txt
pymupdf==1.22.5
langchain==0.0.267
langchain-community==0.0.13
sentence-transformers==2.2.2
huggingface-hub==0.12.1
chromadb==0.4.18
tqdm==4.65.0
openai==1.3.0
pandas==2.0.3
python-dotenv==1.0.0